feat: cleanup-ui-factory-session-detail-test-split-004 - Narrow lint debt and document focused verification
